refactor: root and network routes
