+++ src/routes/root/index.ts
@@ -1,13 +1,10 @@
-import { MetricsCollector } from '@blockfrost/blockfrost-utils';
-import { FastifyInstance } from 'fastify';
-import { getConfig } from '../../config';
 import { getSchemaForEndpoint } from '@blockfrost/openapi';
+import { FastifyInstance } from 'fastify';
+
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const packageJson = require('../../../package.json');
 
-const config = getConfig();
-
-async function root(fastify: FastifyInstance) {
+async function route(fastify: FastifyInstance) {
   fastify.route({
     url: '/',
     method: 'GET',
@@ -18,19 +15,6 @@ async function root(fastify: FastifyInstance) {
         version: `${packageJson.version}`,
       }),
   });
-
-  if (config.server.prometheusMetrics) {
-    const metricsCollector = new MetricsCollector(10_000, {
-      prefix: `blockfrost_backend_ryo`,
-    });
-
-    fastify.route({
-      url: '/prometheus',
-      method: 'GET',
-      handler: async (_request, reply) =>
-        reply.header('Content-Type', 'text/plain').send(metricsCollector.toPrometheus()),
-    });
-  }
 }
 
-module.exports = root;
+export default route;
